server/requirements.txt
flask
python-dotenv
huggingface_hub[hf_xet]
transformers
torch
joblib
pandas
